fix(kb): kb_documents 스키마 드리프트 정합 — project_id INT→BIGINT + 누락 FK 2개
운영은 이미 정상이었다 — project_id bigint · project_id/client_id FK(둘 다 ON DELETE SET NULL)
· 고아 0건. **dev DB 만** project_id 가 int 로 남아 projects.id(bigint)와 타입이 안 맞았고,
그래서 sync-database 가 FK 를 만들지 못하고 매번 조용히 실패해 왔다. 스키마 전체에서
project_id 를 가진 22개 테이블이 전부 bigint 인데 kb_documents 하나만 int 였다.
모델도 INTEGER 로 선언돼 있어 **운영 실제 스키마와도 어긋나** 있었다.

- models/KbDocument.js: project_id INTEGER → BIGINT (운영 실제와 정합)
- 신규 scripts/fix-kb-documents-fk-drift.js: 진단/dry-run 기본 + --apply.
  이미 정합이면 "할 일 없음" 으로 no-op — 운영에서 돌려도 변경 0 (멱등)
- dev DB 적용 완료: 고아 2건 project_id → NULL · INT→BIGINT · FK 2개 생성

고아 NULL 처리 근거: 운영 FK 의 DELETE_RULE 이 SET NULL 이라, FK 가 처음부터 있었다면
프로젝트 삭제 시점에 이미 NULL 이 됐을 값이다. 임의 변경이 아니라 뒤늦은 정합 복원.
조회도 `scope='project' AND project_id IN (내 프로젝트)` 라 없는 프로젝트를 가리키는 행은
이미 아무에게도 안 보였다 — 적용 전/후 목록 응답 동일(문서 1건, 고아 미노출)로 실측.

같이 정리(주석만, 로직 0줄): task_actions.js resume() 의 옛 근거 문구("시간 엔진 라운드 재개
경계" → 타임라인 라벨 + revertStatus 의 from_status 파생) · TaskDetailDrawer.tsx:455 의 옛 함수명.

검증(Fable 독립 게이트 VERDICT: PASS):
- FK 실작동 반증 — 없는 project_id 로 INSERT 시 ER_NO_REFERENCED_ROW_2 거부
- SET NULL 실동작 — 임시 프로젝트 삭제 시 문서는 생존하고 project_id 만 NULL (CASCADE 아님)
- 정상 입력 무회귀 — project scope 문서 생성 201, FK 가 막지 않음
- 멀티테넌트 403 · 운영 옛 데이터 sample 회귀 없음 · 직전 사이클 무회귀(hold/resume 10ms)
- sync-database EXIT 0 / 실 에러 0 / FK 재생성 후에도 SET NULL 유지 → dev = 운영
- 가드 22/22 · health-check 33/33 · tenant e2e 0 · 빌드 exit 0 / error TS 0

운영 영향: **DB 마이그레이션 불필요.** 모델 파일 배포 + pm2 restart 만.

+++ dev-backend/models/KbDocument.js
@@ -102,7 +102,10 @@ KbDocument.init({
     defaultValue: 'workspace',
   },
   project_id: {
-    type: DataTypes.INTEGER,
+    // ★ BIGINT — projects.id 가 bigint 다. INTEGER 로 선언돼 있던 탓에 dev 에서 FK 생성이 매번
+    //   실패했고(타입 불일치), 모델이 운영 실제 스키마와도 어긋나 있었다(운영은 bigint + FK 정상).
+    //   2026-07-31 정합. 다른 project_id FK 22개는 전부 bigint 였고 이 컬럼만 int 였다.
+    type: DataTypes.BIGINT,
     allowNull: true,
     references: { model: 'projects', key: 'id' },
   },
